count and radix sort
Uploading two new ones. Radix uses Count inside of it, so that's why I made Count to begin with. When I started pulling out the strings for the integers I knew it was game time - super fun one. This algo only workw with integer numbers less than 1 billion (due to how I'm figuring out the length of an integer), so if I'm working with huge numbers, time to move to quick or heap haha.