Rename radixSortLSD.pde to radixSortLSDslow.pde
After consulting my teacher, this way is super inefficient! Having all this data being constantly copied between places was super painful for the computer.  I'm going to keep this here because I still think it was a super cool point of using Strings in my sort, but I'm going to create another radix LSD sort that is a little bit more casual than that.